+++ app/src/main/java/org/pathfinderfr/app/character/CharacterSheetActivity.java
@@ -167,7 +167,10 @@ public class CharacterSheetActivity extends AppCompatActivity {
             getIntent().putExtra(CharacterSheetActivity.SELECTED_CHARACTER_ID, characterId);
         }
 
-        String baseText = getResources().getString(R.string.sheet_menu_activity) + " - ";
+        String baseText = (character == null || character.getName() == null || character.getName().length() == 0 ?
+                getResources().getString(R.string.sheet_menu_activity):
+                character.getName());
+        baseText += " - ";
         setTitle(baseText + getResources().getString(R.string.sheet_menu_main));
 
         BottomNavigationView navigation = (BottomNavigationView) findViewById(R.id.sheet_navigation);
@@ -183,7 +186,6 @@ public class CharacterSheetActivity extends AppCompatActivity {
         // no state to be restored
         if(savedInstanceState != null) {
             Log.i(CharacterSheetActivity.class.getSimpleName(), "onCreate with savedInstanceState!");
-            showTab(); // refresh
         } else {
             // initialize tab based on preferences
             switch(currentTab) {