+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Conditions, weapons and armors 
+- Character inventory
 
 ## [2.4.1] - 2019-04-09
 ### Fixed